+++ src/components/ui/DialogDelete.tsx
@@ -1,5 +1,5 @@
 import { MdOutlineDeleteOutline } from "react-icons/md";
-import { Button } from "./@/components/ui/button";
+import { Button } from "../@/components/ui/button";
 import {
   Dialog,
   DialogClose,
@@ -9,7 +9,7 @@ import {
   DialogHeader,
   DialogTitle,
   DialogTrigger,
-} from "./@/components/ui/dialog";
+} from "../@/components/ui/dialog";
 import type { DialogDeleteProps } from "src/types";
 
 function DialogDelete({
