+++ src/components/ui/ToggleTheme.tsx
@@ -1,7 +1,7 @@
-import { useTheme } from "../hooks/useTheme";
+import { useTheme } from "../../hooks/useTheme";
 
 import { HiOutlineMoon, HiOutlineSun } from "react-icons/hi2";
-import { Button } from "./@/components/ui/button";
+import { Button } from "../@/components/ui/button";
 
 function DarkModeToggle() {
   const { isDarkMode, toggleDarkMode } = useTheme();
